+++ app/config/routing.yml
@@ -1,8 +1,3 @@
-wallabag_import:
-    resource: "@WallabagImportBundle/Controller/"
-    type: annotation
-    prefix: /import
-
 app:
     resource: "@WallabagCoreBundle/Controller/"
     type: annotation
