+++ src/Wallabag/CoreBundle/Controller/Import/WallabagV1Controller.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Wallabag\ImportBundle\Controller;
+namespace Wallabag\CoreBundle\Controller\Import;
 
 use Craue\ConfigBundle\Util\Config;
 use OldSound\RabbitMqBundle\RabbitMq\Producer as RabbitMqProducer;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
-use Wallabag\ImportBundle\Import\WallabagV2Import;
+use Wallabag\ImportBundle\Import\WallabagV1Import;
 use Wallabag\ImportBundle\Redis\Producer as RedisProducer;
 
-class WallabagV2Controller extends WallabagController
+class WallabagV1Controller extends WallabagController
 {
-    private WallabagV2Import $wallabagImport;
+    private WallabagV1Import $wallabagImport;
     private Config $craueConfig;
     private RabbitMqProducer $rabbitMqProducer;
     private RedisProducer $redisProducer;
 
-    public function __construct(WallabagV2Import $wallabagImport, Config $craueConfig, RabbitMqProducer $rabbitMqProducer, RedisProducer $redisProducer)
+    public function __construct(WallabagV1Import $wallabagImport, Config $craueConfig, RabbitMqProducer $rabbitMqProducer, RedisProducer $redisProducer)
     {
         $this->wallabagImport = $wallabagImport;
         $this->craueConfig = $craueConfig;
@@ -26,7 +26,7 @@ class WallabagV2Controller extends WallabagController
     }
 
     /**
-     * @Route("/wallabag-v2", name="import_wallabag_v2")
+     * @Route("/import/wallabag-v1", name="import_wallabag_v1")
      */
     public function indexAction(Request $request, TranslatorInterface $translator)
     {
@@ -46,6 +46,6 @@ class WallabagV2Controller extends WallabagController
 
     protected function getImportTemplate()
     {
-        return '@WallabagImport/WallabagV2/index.html.twig';
+        return '@WallabagImport/WallabagV1/index.html.twig';
     }
 }
