+++ src/Wallabag/CoreBundle/Controller/Import/WallabagV2Controller.php
@@ -1,32 +1,32 @@
 <?php
 
-namespace Wallabag\ImportBundle\Controller;
+namespace Wallabag\CoreBundle\Controller\Import;
 
 use Craue\ConfigBundle\Util\Config;
 use OldSound\RabbitMqBundle\RabbitMq\Producer as RabbitMqProducer;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
-use Wallabag\ImportBundle\Import\ShaarliImport;
+use Wallabag\ImportBundle\Import\WallabagV2Import;
 use Wallabag\ImportBundle\Redis\Producer as RedisProducer;
 
-class ShaarliController extends HtmlController
+class WallabagV2Controller extends WallabagController
 {
-    private ShaarliImport $shaarliImport;
+    private WallabagV2Import $wallabagImport;
     private Config $craueConfig;
     private RabbitMqProducer $rabbitMqProducer;
     private RedisProducer $redisProducer;
 
-    public function __construct(ShaarliImport $shaarliImport, Config $craueConfig, RabbitMqProducer $rabbitMqProducer, RedisProducer $redisProducer)
+    public function __construct(WallabagV2Import $wallabagImport, Config $craueConfig, RabbitMqProducer $rabbitMqProducer, RedisProducer $redisProducer)
     {
-        $this->shaarliImport = $shaarliImport;
+        $this->wallabagImport = $wallabagImport;
         $this->craueConfig = $craueConfig;
         $this->rabbitMqProducer = $rabbitMqProducer;
         $this->redisProducer = $redisProducer;
     }
 
     /**
-     * @Route("/shaarli", name="import_shaarli")
+     * @Route("/import/wallabag-v2", name="import_wallabag_v2")
      */
     public function indexAction(Request $request, TranslatorInterface $translator)
     {
@@ -36,16 +36,16 @@ class ShaarliController extends HtmlController
     protected function getImportService()
     {
         if ($this->craueConfig->get('import_with_rabbitmq')) {
-            $this->shaarliImport->setProducer($this->rabbitMqProducer);
+            $this->wallabagImport->setProducer($this->rabbitMqProducer);
         } elseif ($this->craueConfig->get('import_with_redis')) {
-            $this->shaarliImport->setProducer($this->redisProducer);
+            $this->wallabagImport->setProducer($this->redisProducer);
         }
 
-        return $this->shaarliImport;
+        return $this->wallabagImport;
     }
 
     protected function getImportTemplate()
     {
-        return '@WallabagImport/Shaarli/index.html.twig';
+        return '@WallabagImport/WallabagV2/index.html.twig';
     }
 }
