+++ tests/Wallabag/CoreBundle/Controller/Import/ImportControllerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Wallabag\ImportBundle\Controller;
+namespace Tests\Wallabag\CoreBundle\Controller\Import;
 
 use Tests\Wallabag\CoreBundle\WallabagCoreTestCase;
