+++ tests/Wallabag/CoreBundle/fixtures/Import/firefox-bookmarks.json
