+++ tests/Wallabag/CoreBundle/fixtures/Import/unnamed.png
